fix: 修复表格列无法拖拽的问题 (#157)
* fix: 添加修改

* fix: 调整流水线

* fix: 调整流水线

* fix: 修复校验逻辑

* fix: 修复校验逻辑

* feat: 添加正式版本的ci执行

* feat: 查看当前的release的ci

* fix: 修复表格列不显示的问题

* fix: 删除多余的引用

* fix: 删除多余的引用
