+++ packages/react-ui/docs/rwp/changelog/react-ui.md
@@ -45,7 +45,13 @@ group:
 
 > (*/ω＼*) 是不是有些小贪心了。 (/ω＼*)……… (/ω•＼*)
 
-## alpha 内测开发版本
+## canary 内测开发版本
+
+### 1.21.5-canary
+
+- Table
+  - 🐞 修复表格列拖拽的后不显示的问题
+
 
 ### 1.22.2-alpha.0
