+++ packages/react-ui/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@rwp/react-ui",
-    "version": "1.21.4-alpha.0",
+    "version": "1.21.5-canary",
     "license": "AGPL-3.0",
     "scripts": {
         "start": "dumi dev",
