+++ packages/react-ui/src/table/plugin/style/tools.less
@@ -4,6 +4,7 @@
 
 .@{table-prefix-cls}-right{
     display: flex;
+    position: relative;
     border: 1px solid #ddd;
     background-color: #f9f9f9;
     width: 20px;
@@ -22,27 +23,29 @@
             writing-mode: vertical-rl;
        }
     }
-    .@{table-prefix-cls}-right-content {
-        position: absolute;
-        z-index: 1000;
-        border-right: 1px solid #ddd;
-        border-left: 1px solid #ddd;
-        // border-bottom: 1px solid #ddd;
-        width: 250px;
-        margin-left: -250px;
-        background: #f9f9f9;
-        height: calc(100% - 35px) ;
+
+    .@{table-prefix-cls}-right-group {
+        height: 100px;
+        width: 100%;
+        border-top: 1px solid #ddd;
     }
+}
+
+.@{table-prefix-cls}-right-content {
+    z-index: 100;
+    border: 1px solid #ddd;
+    width: 250px;
+    height: 100%;
     .@{table-prefix-cls}-right-column {
         >h4{
             height: 28px;
             margin-bottom: 0px;
             margin-left: 1rem;
+            background: #f9f9f9;
             display: flex;
             align-items: center; 
         }
         >.ant-tree{
-            border-bottom: 1px solid #ddd;
             border-top: 1px solid #ddd;
             .drag-over{
                 color: white;
@@ -53,11 +56,5 @@
                 }
             }
         }
-        
-    }
-    .@{table-prefix-cls}-right-group {
-        height: 100px;
-        width: 100%;
-        border-top: 1px solid #ddd;
     }
 }
